% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{generate_sites}
\alias{generate_sites}
\title{Generate sites with weighted patient distribution}
\usage{
generate_sites(
  raw_site_df,
  sampled_subj,
  new_study_id,
  target_site_count,
  vSiteIDs
)
}
\arguments{
\item{raw_site_df}{Raw site data frame}

\item{sampled_subj}{Data frame of sampled subjects}

\item{new_study_id}{New study ID string}

\item{target_site_count}{Target number of sites to generate}

\item{vSiteIDs}{Site ID column names in hierarchical order}
}
\value{
List with generated_sites and updated_sampled_subj
}
\description{
Generate sites with weighted patient distribution
}
\keyword{internal}
