% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.studykri-package.R
\docType{package}
\name{gsm.studykri-package}
\alias{gsm.studykri}
\alias{gsm.studykri-package}
\title{gsm.studykri: Bootstrap Study KRI comnfidence intervals based on reference studies}
\description{
A new approach for calculating KRI lower and upper limits on study-level for quality monitoring in clinical trials. The method uses bootstrapping to calculate confidence intervals for a given study over-time. The confidence intervals can then be used to compare the study-level KRI against a fixed expectation or against the confidence intervals and KRI values over time of one or more reference studies. The bootstrapping method resamples a new set of sites with replacement from the original study data set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impala-consortium.github.io/gsm.studykri/}
}

}
\author{
\strong{Maintainer}: Bjoern Koneswarakantha \email{bjoern.koneswarakantha@roche.com} (\href{https://orcid.org/0000-0003-4585-7799}{ORCID})

}
\keyword{internal}
