% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{process_derived_domain}
\alias{process_derived_domain}
\title{Process derived domain (uses subject_nsv for lookup)}
\usage{
process_derived_domain(df, subject_mapping, new_study_id, raw_subj)
}
\description{
Process derived domain (uses subject_nsv for lookup)
}
\keyword{internal}
