% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{process_site_domain}
\alias{process_site_domain}
\title{Process site metadata domain}
\usage{
process_site_domain(df, new_study_id, used_site_ids)
}
\description{
Process site metadata domain
}
\keyword{internal}
