% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{process_study_domain}
\alias{process_study_domain}
\title{Process study metadata domain}
\usage{
process_study_domain(df, new_study_id)
}
\description{
Process study metadata domain
}
\keyword{internal}
