% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{process_subject_domain}
\alias{process_subject_domain}
\title{Process subject-level domain}
\usage{
process_subject_domain(df, subject_mapping, new_study_id, sampled_subj)
}
\description{
Process subject-level domain
}
\keyword{internal}
