% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-HandleLazyTable.R
\name{SortDf}
\alias{SortDf}
\title{Sort Data Frame or Lazy Table}
\usage{
SortDf(data, ...)
}
\arguments{
\item{data}{data.frame or tbl_lazy}

\item{...}{Columns to sort by}
}
\value{
Sorted data
}
\description{
Sort Data Frame or Lazy Table
}
\keyword{internal}
