% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePortfolio.R
\name{update_composite_id}
\alias{update_composite_id}
\title{Update composite IDs by replacing old subjid with new subjid}
\usage{
update_composite_id(composite_id, old_subjid, new_subjid)
}
\description{
Update composite IDs by replacing old subjid with new subjid
}
\keyword{internal}
