% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.studykri-package.R
\docType{package}
\name{gsm.studykri-package}
\alias{gsm.studykri}
\alias{gsm.studykri-package}
\title{gsm.studykri: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impala-consortium.github.io/gsm.studykri/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
